+++ settings.json
@@ -1,1 +1,1 @@
-{"device": "cpu", "dt": 11, "file_fmt": ".bmp", "folder": "C:/Users/LabPC/Desktop/exp", "iter_scale": 2.0, "iterations": 2, "overlap": 32, "regime": "offline", "resize": 2, "save_dir": "C:/Users/LabPC/Desktop/Work/TorchPIV/Out", "save_opt": "Save statistics", "scale": 0.0165364, "wind_size": 64}
+{"device": "NVIDIA GeForce GTX 1050 Ti", "dt": 11, "file_fmt": ".bmp", "folder": "C:/Users/LabPC/Desktop/exp", "iter_scale": 2.0, "iterations": 2, "overlap": 32, "regime": "offline", "resize": 2, "save_dir": "C:/Users/LabPC/Desktop/Work/TorchPIV/Out", "save_opt": "Save statistics", "scale": 0.0165364, "wind_size": 64}
